+++ src/Products/PloneMeeting/adapters.py
@@ -1881,9 +1881,11 @@ class IconifiedCategoryGroupAdapter(object):
             return cfg.annexes_types.meeting_annexes
 
     def get_every_categories(self, only_enabled=True):
-        categories = get_categories(self.context, only_enabled=only_enabled)
+        categories = get_categories(
+            self.context, the_objects=True, only_enabled=only_enabled)
         self.request['force_use_item_decision_annexes_group'] = True
-        categories = categories + get_categories(self.context, only_enabled=only_enabled)
+        categories = categories + get_categories(
+            self.context, the_objects=True, only_enabled=only_enabled)
         self.request['force_use_item_decision_annexes_group'] = False
         return categories