+++ src/Products/PloneMeeting/migrations/migrate_to_4204.py
@@ -61,6 +61,7 @@ class Migrate_To_4204(Migrator):
             _configurePortalRepository()
             self._reloadItemTemplateAndRecurringTypes()
             self._initMeetingsItemAttendeesOrder()
+            self.updateHolidays()
         self._migrateMCYearlyInitMeetingNumber()
 
         # ToolPloneMeeting.get_plone_groups_for_user parameter userId is now user_id