+++ CHANGES.rst
@@ -10,7 +10,7 @@ Changelog
   defined on the contacts.
   [gbastien]
 - Escape annex and annex_type title in `ContainedAnnexesVocabulary` and
-  `ContainedDecisionAnnexesVocabulary` in case it contain malicious code.
+  `ContainedDecisionAnnexesVocabulary` in case it contains malicious code.
   [gbastien]
 - Fixed bug where order of annexes of an item sent to another MC was not correct.
   This relies on a change in `collective.iconifiedcategory`,